pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "style-transfer"
version = "0.1.0"
description = "Fast neural style transfer desktop application"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.3.0",
    "torchvision>=0.18.0",
    "onnxruntime-directml>=1.18.0",
    "Pillow>=10.3.0",
    "opencv-python>=4.9.0",
    "PySide6>=6.7.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-qt>=4.4",
    "pytest-mock>=3.14",
    "pytest-cov>=5.0",
    "ruff",
    "mypy",
    "tensorboard",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short -v"
